fix: カスタムフィールドの重複登録修正

+++ src/System/Helper/WpCustomHelper.php
@@ -6,6 +6,10 @@ use System\Models\CustomPostType;
 
 final class WpCustomHelper
 {
+    private static array $registered_posts = [];
+
+    private static array $registered_fields = [];
+
     public static function addCustomPostType(string $class_name): void
     {
         $post_slug = self::addPostType($class_name);
@@ -16,15 +20,18 @@ final class WpCustomHelper
     {
         $slug = CustomPostType::getSlug($class_name);
         $title = CustomPostType::getTitle($class_name);
-        add_action('init', function () use ($slug, $title) {
-            register_post_type($slug, [
-                'label' => $title,
-                'public' => true,
-                'has_archive' => false,
-                'menu_position' => 5,
-                'show_in_rest' => false,
-            ]);
-        });
+        if (!array_key_exists($slug, self::$registered_posts)) {
+            add_action('init', function () use ($slug, $title) {
+                register_post_type($slug, [
+                    'label' => $title,
+                    'public' => true,
+                    'has_archive' => false,
+                    'menu_position' => 5,
+                    'show_in_rest' => false,
+                ]);
+            });
+            self::$registered_posts[$slug] = true;
+        }
 
         return $slug;
     }
@@ -43,44 +50,47 @@ final class WpCustomHelper
     {
         $slug = $post_slug . '_' . $field_slug;
         $nonce = '_wp_nonce_' . $field_slug;
-        $key = $field_slug . '_key';
         $id = $field_slug . '_id';
         $name = $field_slug . '_name';
-        add_action('add_meta_boxes', function () use ($slug, $nonce, $key, $id, $name, $post_slug, $field_slug, $title) {
-            add_meta_box(
-                $slug,
-                $title,
-                function () use ($nonce, $key, $id, $name, $field_slug) {
-                    $value = esc_attr(get_post_meta(get_the_ID(), $field_slug, true));
-                    wp_nonce_field('wp-nonce-key', $nonce);
-                    echo "<div class=\"{$field_slug}\">";
-                    echo "<input type=\"text\" value=\"{$value}\" id=\"{$id}\" name=\"{$name}\">";
-                    echo "</div>";
-                },
-                // null,
-                // 'advanced',
-                // 'default',
-            );
-        });
+        if (!array_key_exists($slug, self::$registered_fields)) {
+            add_action('add_meta_boxes', function () use ($slug, $post_slug, $nonce, $id, $name, $field_slug, $title) {
+                add_meta_box(
+                    $slug,
+                    $title,
+                    function () use ($nonce, $id, $name, $field_slug) {
+                        $value = esc_attr(get_post_meta(get_the_ID(), $field_slug, true));
+                        wp_nonce_field('wp-nonce-key', $nonce);
+                        echo "<div class=\"{$field_slug}\">";
+                        echo "<input type=\"text\" value=\"{$value}\" id=\"{$id}\" name=\"{$name}\">";
+                        echo "</div>";
+                    },
+                    screen: $post_slug,
+                    context: 'advanced',
+                    priority: 'default',
+                    callback_args: null,
+                );
+            });
 
-        add_action('save_post', function ($post_id) use ($nonce, $name, $field_slug) {
-            if (!isset($_POST[$nonce]) || !$_POST[$nonce]) {
-                return;
-            }
-            if (!check_admin_referer('wp-nonce-key', $nonce)) {
-                return;
-            }
-            if (!isset($_POST[$name])) {
-                return;
-            }
+            add_action('save_post', function ($post_id) use ($nonce, $name, $field_slug) {
+                if (!isset($_POST[$nonce]) || !$_POST[$nonce]) {
+                    return;
+                }
+                if (!check_admin_referer('wp-nonce-key', $nonce)) {
+                    return;
+                }
+                if (!isset($_POST[$name])) {
+                    return;
+                }
 
-            $value = $_POST[$name];
-            if (!empty($value)) {
-                add_post_meta($post_id, $field_slug, $value, true);
-                update_post_meta($post_id, $field_slug, $value);
-            } elseif ($value === '') {
-                delete_post_meta($post_id, $field_slug);
-            }
-        });
+                $value = $_POST[$name];
+                if (!empty($value)) {
+                    add_post_meta($post_id, $field_slug, $value, true);
+                    update_post_meta($post_id, $field_slug, $value);
+                } elseif ($value === '') {
+                    delete_post_meta($post_id, $field_slug);
+                }
+            });
+            self::$registered_fields[$slug] = true;
+        }
     }
 }
